% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUniFrac.R
\name{LUniFrac}
\alias{LUniFrac}
\title{LUniFrac}
\usage{
LUniFrac(otu.tab, tree, alpha = c(0, 0.5, 1), metadata)
}
\arguments{
\item{otu.tab}{OTU count table, containing 2*n rows (samples) and q columns (OTUs)}

\item{tree}{Rooted phylogenetic tree of R class "phylo"}

\item{alpha}{Parameter controlling weight on abundant lineages. The same weight is used within a subjects as between subjects.}

\item{metadata}{Data frame with three columns: sample identifiers (must match row names of otu.tab), subject identifiers (n unique values), and time point (variable with two unique levels).}
}
\value{
Returns a (K+1) dimensional array containing the longitudinal UniFrac dissimilarities with the K specified alpha values plus the unweighted distance.
}
\description{
Longitudinal UniFrac distances for comparing changes in
microbial communities across 2 time points.
}
\details{
Based in part on Jun Chen & Hongzhe Li (2012), GUniFrac.

Computes difference between time points and then calculates
difference of these differences, resulting in a dissimilarity
matrix that can be used as-is or converted to a kernel
(similarity) matrix for hypothesis testing with MIRKAT
and related methods.
}
